+++ NEWS.rst
@@ -15,6 +15,8 @@ bundle_adjust (:numref:`bundle_adjust`):
 mapproject (:numref:`mapproject`):
   * If the input DEM is in the ``longlat`` projection, a projection 
     in meters is auto-determined (:numref:`mapproj_auto_proj`).
+  * Added the option ``--ref-map`` to borrow the grid size and projection from
+    an existing mapprojected image (:numref:`mapproj_refmap`).
   * Add the option ``--query-pixel``.
 
 jitter_solve (:numref:`jitter_solve`):